cmd: split auth files and extract context setup
Address review feedback:
- Split auth.go into auth.go (group) + auth_login.go (login command)
- Extract cmdContext, skipAuth, initContext into context.go
- Soften help text wording: "instead" → "also"
